Refactor about page and fix dimensions
In about/page.tsx, raw elements replaced with Hero and LegalDoc components which were broken out of the module.

Firstly, the Hero section is exported to a new functional component in Hero.tsx. This section contains the main heading of the page and its design.
Secondly, several <Link> tags are replaced by the functional component LegalDoc.tsx in order to reduce redundancies. This is possible as the structure of the legal documents was similar. As a result, replacing the repeated code with a reusable LegalDoc component that accepts props for individual attributes simplifies the codebase and makes it more scalable.

Besides the described changes, removes unnecessary imports for about/page.tsx.
Notice: no visual changes should apply, all elements will remain visually the same. Changes related to code cleanliness and organization only.